// enlista-reels/src/components/animations/TextReveal.d.ts
import React from 'react';
interface TextRevealProps {
    text: string;
    startFrame: number;
    endFrame: number;
    fontSize: number;
    fontWeight: number;
    color: string;
    fontFamily: string;
}
export declare const TextReveal: React.FC<TextRevealProps>;
export {};
//# sourceMappingURL=TextReveal.d.ts.map