// enlista-reels/src/components/animations/CardSlideIn.d.ts
import React from 'react';
interface CardSlideInProps {
    children: React.ReactNode;
    startFrame: number;
    endFrame: number;
    direction: 'left' | 'right' | 'top' | 'bottom';
    distance: number;
}
export declare const CardSlideIn: React.FC<CardSlideInProps>;
export {};
//# sourceMappingURL=CardSlideIn.d.ts.map